pyproject.toml
[project]
name = "dendrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-snap7>=2.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]
